resume_builder/requirements.txt
fastapi
uvicorn
firebase-admin
pylatex
python-dotenv